% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{calculateMeanValues}
\alias{calculateMeanValues}
\title{Calculate the mean intensity for each trace and return it in a matrix format}
\usage{
calculateMeanValues(fiaExp)
}
\arguments{
\item{fiaExp}{the xcmsExp object for one fia batch}
}
\value{
a tibble with three columns, fName, sName and fiaValue.
fName is the feature (mrm) name, sName is the sample name and fiaValue
is the the fia cps value for a particular sample and feature.
}
\description{
Calculate the mean intensity for each trace and return it in a matrix format
}
\details{
The core function of this package is to calculate a single value for each FIA trace.
That core function is performed here.
}
