% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentClass.R
\name{fakeProgress}
\alias{fakeProgress}
\title{Fake progress}
\usage{
fakeProgress(value = NULL, detail = NULL)
}
\arguments{
\item{value}{double ranging from 0 to 1 indicating the progress}

\item{detail}{text for the progress bar}
}
\description{
Fake progress
}
\details{
This is a fake progress function used when no progress function is passed.
The real progress function comes from shiny and is used to give the user
and indicator of the progress of a long process
}
