% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentClass.R
\name{getFIAApp}
\alias{getFIAApp}
\title{Get the FIA MS SS app}
\usage{
getFIAApp(self)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}
}
\value{
shinyApp for the FIA SS checking
}
\description{
Get the FIA MS SS app
}
