% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentClass.R
\name{getSettingsApp}
\alias{getSettingsApp}
\title{get the shiny app for the settings}
\usage{
getSettingsApp(self)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}
}
\value{
shinyApp for the settings
}
\description{
get the shiny app for the settings
}
