% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentClass.R
\name{initFiaR6}
\alias{initFiaR6}
\title{initialize the fiaR6 object}
\usage{
initFiaR6(self, workdirPath, wiffPath)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}

\item{workdirPath}{the working directory path, contains both mzml data files, the results
and settings (if they are saved)}

\item{wiffPath}{the path to the topmost directory containing raw data files (to be converted via
proteowizard's msconvert)}
}
\value{
invisible(self)
}
\description{
initialize the fiaR6 object
}
