% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{loadFiaResults}
\alias{loadFiaResults}
\title{Load the FIA results}
\usage{
loadFiaResults(self, updateProgress)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}

\item{reload}{Boolean. Should the data be refreshed from result files. Will cause
\code{\link{reloadFiaResults}} to be called if TRUE. Will reload the result files
generated from \code{\link{processOneFolder}}}

\item{forceRecalc}{Boolean. Will cause \code{\link{reloadFiaResults}} to be
called if TRUE. Will also force the recalculation of the results.}
}
\value{
no return value, assigns to self$resdata
and self$resdataNice the loaded values
}
\description{
Load the FIA results
}
\details{
Load the FIA resutls, passes the forceRecalc variable to the underlying
reloadFiaResults function that is called if the RData files are not found
}
