% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{parseDups}
\alias{parseDups}
\title{parse out diplicates from a tibble}
\usage{
parseDups(fileTibble)
}
\arguments{
\item{fileTibble}{a tibble with columns of "barcode",
"well", "runNo", "extra1", "extra2", "sampleType", "sampleName"
coming from the parseFile function}
}
\value{
A tibble with a new colum batchNo indicating the batch number
}
\description{
parse out diplicates from a tibble
}
\details{
parsing a number of filenames where sometimes the same
file name is generated at different times (different batches)
This function adds another column (batchNo) which is a running
number that indicates different batches with the same file name.
}
