% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_load.R
\name{parseFilename}
\alias{parseFilename}
\title{Split a file name by '.' or '_' or '-'}
\usage{
parseFilename(oneName)
}
\arguments{
\item{oneName}{the filename that needs to be parsed}
}
\value{
a vector with the strings split from the file name
}
\description{
Split a file name by '.' or '_' or '-'
}
\details{
Takes a filename ad splits it by one of three characters ('.' or '_' or '-').
TODO: add the return indices to the call
}
