% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{parseFilenames}
\alias{parseFilenames}
\title{Split a file name by '.' or '_' or '-'}
\usage{
parseFilenames(fileTibble, splitcols, sortcols)
}
\arguments{
\item{fileTibble}{a tibble that contains file names in a column called sName}

\item{splitcols}{a character vector containing the column names of the split. Any
column whose name starts with 'XX' will be discarded before returning the tible}

\item{sortcols}{a character vector containing the columns to be used for sorting and
comparison for duplicate runs (basically different batches but same file name)}
}
\value{
a tibble with strings split from the file name and two more columns added
one indicating the run time of the file and the other indicating the batch number
if the information in the tibble is otherwise identical
}
\description{
Split a file name by '.' or '_' or '-'
}
\details{
Every filename and splits it by one of three characters ('.' or '_' or '-').
}
