% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentClass.R
\name{prepForFIA}
\alias{prepForFIA}
\title{Prepare for FIA SS data exploration}
\usage{
prepForFIA(self, updateProgress = NULL)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}
}
\value{
invisible(self)
}
\description{
First this function looks for any new wiff files to add to the results. Then it
tries to load the data from the converted files and put them into data objects
that can be used for data exploration
}
