% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{processOneFolder}
\alias{processOneFolder}
\title{Process one mzML folder}
\usage{
processOneFolder(folderPath, resultName, forceRecalc = FALSE)
}
\arguments{
\item{folderPath}{the path to the mzML folder}
}
\description{
Use multi-core processing (if set-up) to read mzML files from one directory
and write out the results for each file
}
