% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{processOneMZML}
\alias{processOneMZML}
\title{Process one mzML file}
\usage{
processOneMZML(filePath)
}
\arguments{
\item{filePaht}{the path (including the filename) of the mzML file}
}
\value{
Nothing
}
\description{
Read one mzML file, calculate mean values for each transition and write the
result into a tsv table with the same name as the mzML file
}
