% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{readFiadata}
\alias{readFiadata}
\title{Read a collection of fia files, and split them into pos and neg polarities}
\usage{
readFiadata(fiaFolderPath)
}
\arguments{
\item{fiaFolderPath}{a path to a batch of mzML files to be read and process}
}
\value{
a list of possibly one or two xcms objects, one for each polarity containing the
respective mzML file information. The list items are named, either 'pos' and/or 'neg'
reflecting the polarity of the data. This is obviously assuming that one data file only
stores data from one polarity.
}
\description{
Read a collection of fia files, and split them into pos and neg polarities
}
\details{

}
