% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{readOneFolder}
\alias{readOneFolder}
\title{Read results from a folder containing data files from one experiment}
\usage{
readOneFolder(self, onePath, resultName = "result.tsv",
  forceRecalc = FALSE)
}
\arguments{
\item{onePath}{a path to a batch of SS files generated}

\item{resultName}{the name of the file that is used to save the result}

\item{forceRecalc}{boolean. If TRUE then all mzML files are reprocessed.
If FALSE then the already generated results are looked for first and returned
if found.}
}
\value{
a tibble of results for all experiments found in the folder or
an emtpy tibble if no results should be returned
}
\description{
Read results from a folder containing data files from one experiment
}
\details{

}
