% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_load.R
\name{reloadFiaResults}
\alias{reloadFiaResults}
\title{reload the fia results from the working directory}
\usage{
reloadFiaResults(forceRecalc = FALSE)
}
\arguments{
\item{forceRecalc}{boolean. If TRUE then mzML files are read
again and the results generated again.}
}
\value{
nothing
}
\description{
reload the fia results from the working directory
}
\details{
Loads the results from each batch folder
and combine them into two tibbles which are then saved in two
RData objects that are used overall in the package. forceRecalc = TRUE
will result in loading the mzML files and re-assessing the FIA value
for all transitions.
}
