% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{runMainProgram}
\alias{runMainProgram}
\title{Run Main Program}
\usage{
runMainProgram(WorkDirPath = "")
}
\arguments{
\item{WorkDirPath}{he path of the working directory that contains 
the mzML files, the FIA results and the settings for the data exploration.}
}
\description{
Run Main Program
}
