% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{loadFiaResults}
\alias{loadFiaResults}
\title{Load the FIA results}
\usage{
loadFiaResults(self, forceRecalc = FALSE)
}
\arguments{
\item{self}{passed from the pR6 object contains all public variables and functions}

\item{forceRecalc}{boolean. Passed to \code{\link{reloadFiaResults}}.}
}
\value{
no return value, assigns to global values globalResdata
and globalResdataNice
}
\description{
Load the FIA results
}
\details{
Load the FIA resutls, passes the forceRecalc variable to the underlying
reloadFiaResults function that is called if the RData files are not found
}
