% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiaDataFunctions.R
\name{parseDups}
\alias{parseDups}
\title{parse out diplicates from a tibble}
\usage{
parseDups(fileTibble, testCols)
}
\arguments{
\item{fileTibble}{a tibble}

\item{testCols}{a string vector with column names to use
for determining if there are more than one batch.}
}
\value{
A tibble with a new colum batchNo indicating the batch number
}
\description{
parse out diplicates from a tibble
}
\details{
parsing a number of filenames where sometimes the same
file name is generated at different times (different batches)
This function adds another column (batchNo) which is a running
number that indicates different batches with the same file name.
}
